// resources/react/src/types/Loto7Types.ts
export type Loto7Types = {
  id: number;
  times: string;
  event_date: Date;
  per_number_1: number;
  per_number_2: number;
  per_number_3: number;
  per_number_4: number;
  per_number_5: number;
  per_number_6: number;
  per_number_7: number;
  bonus_number_1: number;
  bonus_number_2: number;
};